change to new structure
