+++ apps/server/src/modules/school/controller/dto/index.ts
@@ -1,6 +1,3 @@
-export * from './request/migration.body';
-export * from './request/school.params';
-
-export * from './response/migration.response';
-export * from './response/school-external-tool.response';
-export * from './response/school-external-tool-list.response';
+export * from './migration.body';
+export * from './migration.response';
+export * from './school.params';
