+++ apps/server/src/modules/school/school-api.module.ts
@@ -4,14 +4,10 @@ import { SchoolUc } from './uc/school.uc';
 import { SchoolModule } from './school.module';
 import { SchoolController } from './controller/school.controller';
 import { LoggerModule } from '../../core/logger';
-import { SchoolExternalToolController } from './controller/school-external-tool.controller';
-import { ToolModule } from '../tool';
-import { SchoolExternalToolUc } from './uc/school-external-tool.uc';
-import { SchoolExternalToolMapper } from './controller/mapper/school-external-tool.mapper';
 
 @Module({
-	imports: [SchoolModule, AuthorizationModule, LoggerModule, ToolModule],
-	controllers: [SchoolController, SchoolExternalToolController],
-	providers: [SchoolUc, SchoolExternalToolUc, SchoolExternalToolMapper],
+	imports: [SchoolModule, AuthorizationModule, LoggerModule],
+	controllers: [SchoolController],
+	providers: [SchoolUc],
 })
 export class SchoolApiModule {}
