+++ apps/server/src/modules/tool/controller/dto/index.ts
@@ -8,6 +8,8 @@ export * from './request/lti11-launch.query';
 export * from './request/lti11-tool-config.params';
 export * from './request/oauth2-tool-config.params';
 export * from './request/tool-id.params';
+export * from './request/scope.query';
+export * from './request/id.query';
 
 export * from './response/basic-tool-config.response';
 export * from './response/custom-parameter.response';
@@ -17,3 +19,5 @@ export * from './response/external-tool-search-list.response';
 export * from './response/lti11-launch.response';
 export * from './response/lti11-tool-config.response';
 export * from './response/oauth2-tool-config.response';
+export * from './response/school-external-tool.response';
+export * from './response/school-external-tool-list.response';
