+++ apps/server/src/modules/tool/controller/dto/request/id.query.ts
@@ -1,0 +1,9 @@
+import { ApiProperty } from '@nestjs/swagger';
+import { EntityId } from '@shared/domain';
+import { IsMongoId } from 'class-validator';
+
+export class IdQuery {
+	@IsMongoId()
+	@ApiProperty()
+	id!: EntityId;
+}
