+++ apps/server/src/modules/tool/controller/dto/request/scope.query.ts
@@ -1,0 +1,9 @@
+import { IsEnum } from 'class-validator';
+import { ApiProperty } from '@nestjs/swagger';
+import { ConfigurationScope } from '../../../interface';
+
+export class ScopeQuery {
+	@IsEnum(ConfigurationScope)
+	@ApiProperty({ nullable: false, required: true })
+	scope!: ConfigurationScope;
+}
