+++ apps/server/src/modules/tool/controller/dto/response/school-external-tool-list.response.ts
