+++ apps/server/src/modules/tool/controller/dto/response/school-external-tool.response.ts
