+++ apps/server/src/modules/tool/controller/mapper/external-tool-response.mapper.ts
@@ -14,6 +14,8 @@ import {
 	ExternalToolResponse,
 	Lti11ToolConfigResponse,
 	Oauth2ToolConfigResponse,
+	SchoolExternalToolListResponse,
+	SchoolExternalToolResponse,
 } from '../dto';
 
 const scopeMapping: Record<CustomParameterScope, CustomParameterScopeParams> = {
@@ -91,5 +93,22 @@ export class ExternalToolResponseMapper {
 				isOptional: customParameterDO.isOptional,
 			};
 		});
+	}
+
+	mapExternalToolDOsToSchoolExternalToolListResponse(externalTools: ExternalToolDO[]): SchoolExternalToolListResponse {
+		return new SchoolExternalToolListResponse(this.mapExternalToolDOsToSchoolExternalToolResponses(externalTools));
+	}
+
+	private mapExternalToolDOsToSchoolExternalToolResponses(
+		externalTools: ExternalToolDO[]
+	): SchoolExternalToolResponse[] {
+		return externalTools.map(
+			(tool: ExternalToolDO) =>
+				new SchoolExternalToolResponse({
+					id: tool.id || '',
+					name: tool.name,
+					logoUrl: tool.logoUrl,
+				})
+		);
 	}
 }
