+++ apps/server/src/modules/tool/index.ts
@@ -1,4 +1,2 @@
 export * from './tool.module';
 export * from './tool-api.module';
-export * from './service/external-tool.service';
-export * from './service/school-external-tool.service';
