+++ apps/server/src/modules/tool/interface/configuration-scope.enum.ts
@@ -1,0 +1,3 @@
+export enum ConfigurationScope {
+	SCHOOL = 'school',
+}
