+++ apps/server/src/modules/tool/interface/index.ts
@@ -6,3 +6,4 @@ export * from './lti-privacy-permission.enum';
 export * from './lti-role.enum';
 export * from './token-endpoint-auth-method.enum';
 export * from './tool-config-type.enum';
+export * from './configuration-scope.enum';
